# web/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-throughput competitive statistics and tournament infrastructure for Medal of Honor: Allied Assault.",
        "title": "OpenMOHAA Stats API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@moh-central.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "api.moh-central.net",
    "basePath": "/api/v1",
    "paths": {
...
        "/stats/drilldown/options": {
            "get": {
                "description": "Returns available dimensions for drilling down into a stat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Stats"
                ],
                "summary": "Get drilldown options",
                "parameters": [
                    {
                        "type": "string",
                        "default": "kd",
                        "description": "Stat",
                        "name": "stat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrilldownOptionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/stats/leaderboard/combos": {
            "get": {
                "description": "Get leaderboard for derived combo metrics (run_gun, clutch, consistency)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leaderboards"
                ],
                "summary": "Combo leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric (run_gun, clutch, consistency)",
                        "name": "metric",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Max items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ComboLeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stats/leaderboard/contextual": {
            "get": {
                "description": "Get best players for a specific context (e.g. Best Snipers on Mohdm6)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leaderboards"
                ],
                "summary": "Contextual leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "default": "kd",
                        "description": "Stat to rank by",
                        "name": "stat",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Context dimension (e.g. map, weapon)",
                        "name": "dimension",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Context value (e.g. mohdm6, Thompson)",
                        "name": "value",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Max items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ContextualLeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stats/leaderboard/peak": {
            "get": {
                "description": "Get players who excel in specific time windows (morning, night, weekend)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leaderboards"
                ],
                "summary": "Peak performance leaderboard",
                "parameters": [
                    {
                        "type": "string",
                        "default": "evening",
                        "description": "Time Dimension (morning, afternoon, evening, night, weekend)",
                        "name": "dimension",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Max items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PeakLeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/stats/player/{guid}/combos": {
            "get": {
                "description": "Returns complex cross-referenced stats like best weapon per map, victim/killer patterns, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Stats"
                ],
                "summary": "Get player combo metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player GUID",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ComboMetrics"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stats/player/{guid}/drilldown": {
            "get": {
                "description": "Allows drilling down into a specific stat by a dimension (e.g. K/D by Weapon)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Stats"
                ],
                "summary": "Drill down into player stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player GUID",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "kd",
                        "description": "Stat to analyze",
                        "name": "stat",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "weapon",
                        "description": "Dimension to group by",
                        "name": "dimension",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrillDownResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stats/player/{guid}/drilldown/{dimension}/{value}": {
            "get": {
                "description": "Provides a second level of breakdown (e.g. K/D with Thompson on Mohdm6)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Stats"
                ],
                "summary": "Nested drill down",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player GUID",
                        "name": "guid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Parent Dimension (e.g. weapon)",
                        "name": "dimension",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Parent Value (e.g. Thompson)",
                        "name": "value",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Child Dimension (e.g. map)",
                        "name": "child_dimension",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "kd",
                        "description": "Stat to analyze",
                        "name": "stat",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max items to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrillDownNestedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ComboLeaderboardResponse": {
            "type": "object",
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StatLeaderboardEntry"
                    }
                },
                "metric": {
                    "type": "string"
                }
            }
        },
...
        "models.ContextualLeaderboardResponse": {
            "type": "object",
            "properties": {
                "dimension": {
                    "type": "string"
                },
                "leaders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StatLeaderboardEntry"
                    }
                },
                "stat": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "models.DrilldownOptionsResponse": {
            "type": "object",
            "properties": {
                "dimensions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stat": {
                    "type": "string"
                }
            }
        },
...
        "models.PeakLeaderboardResponse": {
            "type": "object",
            "properties": {
                "dimension": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PeakLeaderboardEntry"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerToken": {
            "type": "apiKey",
            "name": "X-Server-Token",
            "in": "header"
        }
    }
}